Update routing rules
It will look for the first path segment that is a valid Controller then
check if the following path segment is a function of that Controller, if
it is then it will call that function with all the following path
segments as parameters to that function, if it's not a valid function
then all segments after the controller will be considered parameters to
the Controller's index function.

Given the path /nothing/something/action/param1/param2?foo=bar
"nothing" is not a controller and will be ignored.
"something" matches the controller SomethingController (based on the
class prefix and suffix) and therefore that controller will be used.
"action" is a function of SomethingController therefore the resulting
call based on a request to that URI will be:
`SomethingController->action('param1', 'param2')`
To access the GET and POST parameters you can to use
$this->input->get('foo');

For the home page (route '/') the default controller will be used with
the default function.